+++ tests/TestModel.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace SanderVanHooft\PayableRedirect;
+
+use Illuminate\Database\Eloquent\Model;
+use SanderVanHooft\PayableRedirect\Payment;
+
+class TestModel extends Model
+{
+    protected $guarded = [];
+
+    public function payments()
+    {
+        return $this->morphMany(Payment::class, 'payable');
+    }
+}